Fixup test failure with delete-from query on MySQL
MySQL doesn't allow queries of the form

delete from <table> where id = select id from <table>

PostgreSQL and H2 allow this, but MySQL will failure to execute the
query. The fix make the query an aliased FROM and wrap it in a
select *. This forces it to a temp table which MySQL allows.