Add support for sending alert emails via a background thread
Previously notifications of alert lifecycles (i.e. subscribing to an
alert, or being removed from an alert) was happening on the HTTP
request thread. When the SMTP server was slow, this would slow down
the response and thus the UI. This instead returns immediately and
connects to the SMTP server via a background thread separately from
the HTTP request thread.